faet: add check for existing context in some ops

+++ src/cmds/stop.go
@@ -3,7 +3,7 @@ package cmds
 import (
 	"fmt"
 
-	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
+	"github.com/Layr-Labs/avs-devnet/src/kurtosis"
 	"github.com/urfave/cli/v2"
 )
 
@@ -13,13 +13,16 @@ func Stop(ctx *cli.Context) error {
 		return cli.Exit(err, 1)
 	}
 	fmt.Println("Stopping devnet...")
-	kurtosisCtx, err := kurtosis_context.NewKurtosisContextFromLocalEngine()
+	kurtosisCtx, err := kurtosis.InitKurtosisContext()
 	if err != nil {
 		return cli.Exit(err, 2)
 	}
+	if !kurtosisCtx.EnclaveExists(ctx.Context, devnetName) {
+		return cli.Exit("Failed to find devnet '"+devnetName+"'. Maybe it's not running?", 3)
+	}
 	if err = kurtosisCtx.DestroyEnclave(ctx.Context, devnetName); err != nil {
-		return cli.Exit(err, 3)
+		return cli.Exit(err, 4)
 	}
 	fmt.Println("Devnet stopped!")
-	return err
+	return nil
 }
